Split voting into list (/vote) and detail (/vote/:pollId) pages
- /vote: poll list with links to individual polls, generic text
- /vote/:pollId: single poll voting page with full card layout
- Track voted option names/icons in localStorage, show in confirmation UI
- Add PollVotePage.tsx, rewrite VotePage.tsx as poll list
- Update Header to support /vote/:pollId pattern route
- Update ESPN links to point to /vote
