+++ src/App.tsx
@@ -40,6 +40,7 @@ import { DonationQueryClientProvider } from './context/donationQueryClientContex
 import { MailingListQueryClientProvider } from './context/mailingListQueryClientContext';
 import { VotingQueryClientProvider } from './context/votingQueryClientContext';
 import VotePage from './pages/VotePage';
+import PollVotePage from './pages/PollVotePage';
 import ErrorBoundary from './components/ErrorBoundary';
 import ConsentBanner from './components/ConsentBanner';
 import { ConsentProvider } from './context/consentContext';
@@ -145,8 +146,12 @@ const App = () => {
                                       element={<ESPNPage />}
                                     />
                                     <Route
-                                      path="/espn/vote"
+                                      path="/vote"
                                       element={<VotePage />}
+                                    />
+                                    <Route
+                                      path="/vote/:pollId"
+                                      element={<PollVotePage />}
                                     />
                                   </Routes>
                                 </main>
